Add files via upload
This is a Machine Learning mini project. It's the 2nd part of the whole project. The dataset is QSAR Biodegradation. In this project we created an adobe spark page and code using python. 
So here we built two predictive models (Neural Network and Fuzzy Logic System) to predict the target variable of the dataset. 
And performed a comparison between the two predictive models.
Reported the accuracy, recall, precision and F1-score measures as well as the confusion matrix.